backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0  # Replace with the actual version
python-dotenv==1.0.0  # Replace with the actual version
requests==2.31.0  # Replace with the actual version
